move roles to Auth context
