+++ lib/elephant_in_the_room/sites/sites.ex
@@ -5,7 +5,7 @@ defmodule ElephantInTheRoom.Sites do
 
   import Ecto.Query, warn: false
   alias ElephantInTheRoom.Repo
-  alias ElephantInTheRoom.Sites.{Site, Category, Post, Tag, Role, Author}
+  alias ElephantInTheRoom.Sites.{Site, Category, Post, Tag, Author}
 
   @doc """
   Returns the list of sites.
@@ -460,102 +460,6 @@ defmodule ElephantInTheRoom.Sites do
   """
   def change_tag(%Tag{} = tag) do
     Tag.changeset(tag, %{})
-  end
-
-  @doc """
-  Returns the list of roles.
-
-  ## Examples
-
-      iex> list_roles()
-      [%Role{}, ...]
-
-  """
-  def list_roles do
-    Repo.all(Role)
-  end
-
-  @doc """
-  Gets a single role.
-
-  Raises `Ecto.NoResultsError` if the Role does not exist.
-
-  ## Examples
-
-      iex> get_role!(123)
-      %Role{}
-
-      iex> get_role!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_role!(id) do
-    Repo.get!(Role, id)
-  end
-
-  @doc """
-  Creates a role.
-
-  ## Examples
-
-      iex> create_role(%{field: value})
-      {:ok, %Role{}}
-
-      iex> create_role(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_role(attrs \\ %{}) do
-    %Role{}
-    |> Role.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a role.
-
-  ## Examples
-
-      iex> update_role(role, %{field: new_value})
-      {:ok, %Role{}}
-
-      iex> update_role(role, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_role(%Role{} = role, attrs) do
-    role
-    |> Role.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a Role.
-
-  ## Examples
-
-      iex> delete_role(role)
-      {:ok, %Role{}}
-
-      iex> delete_role(role)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_role(%Role{} = role) do
-    Repo.delete(role)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking role changes.
-
-  ## Examples
-
-      iex> change_role(role)
-      %Ecto.Changeset{source: %Role{}}
-
-  """
-  def change_role(%Role{} = role) do
-    Role.changeset(role, %{})
   end
 
   @doc """
