+++ lib/elephant_in_the_room_web/controllers/role_controller.ex
@@ -1,57 +1,59 @@
 defmodule ElephantInTheRoomWeb.RoleController do
   use ElephantInTheRoomWeb, :controller
 
-  alias ElephantInTheRoom.Sites
-  alias ElephantInTheRoom.Sites.Role
+  alias ElephantInTheRoom.Auth
+  alias ElephantInTheRoom.Auth.Role
 
   def index(conn, _params) do
-    roles = Sites.list_roles()
+    roles = Auth.list_roles()
     render(conn, "index.html", roles: roles)
   end
 
   def new(conn, _params) do
-    changeset = Sites.change_role(%Role{})
+    changeset = Auth.change_role(%Role{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"role" => role_params}) do
-    case Sites.create_role(role_params) do
+    case Auth.create_role(role_params) do
       {:ok, role} ->
         conn
         |> put_flash(:info, "Role created successfully.")
         |> redirect(to: role_path(conn, :show, role))
+
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, "new.html", changeset: changeset)
     end
   end
 
   def show(conn, %{"id" => id}) do
-    role = Sites.get_role!(id)
+    role = Auth.get_role!(id)
     render(conn, "show.html", role: role)
   end
 
   def edit(conn, %{"id" => id}) do
-    role = Sites.get_role!(id)
-    changeset = Sites.change_role(role)
+    role = Auth.get_role!(id)
+    changeset = Auth.change_role(role)
     render(conn, "edit.html", role: role, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "role" => role_params}) do
-    role = Sites.get_role!(id)
+    role = Auth.get_role!(id)
 
-    case Sites.update_role(role, role_params) do
+    case Auth.update_role(role, role_params) do
       {:ok, role} ->
         conn
         |> put_flash(:info, "Role updated successfully.")
         |> redirect(to: role_path(conn, :show, role))
+
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, "edit.html", role: role, changeset: changeset)
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    role = Sites.get_role!(id)
-    {:ok, _role} = Sites.delete_role(role)
+    role = Auth.get_role!(id)
+    {:ok, _role} = Auth.delete_role(role)
 
     conn
     |> put_flash(:info, "Role deleted successfully.")
