+++ lib/elephant_in_the_room_web/templates/user/form.html.eex
@@ -12,6 +12,12 @@
   </div>
 
   <div class="form-group">
+    <%= label f, :role, class: "control-label" %>
+    <%= select f, :role_id, get_roles() %>
+    <%= error_tag f, :role_id %>
+  </div>
+
+  <div class="form-group">
     <%= label f, :firstname, class: "control-label" %>
     <%= text_input f, :firstname, class: "form-control" %>
     <%= error_tag f, :firstname %>
