+++ lib/elephant_in_the_room_web/views/user_view.ex
@@ -1,3 +1,10 @@
 defmodule ElephantInTheRoomWeb.UserView do
   use ElephantInTheRoomWeb, :view
+
+  alias ElephantInTheRoom.Auth
+
+  def get_roles() do
+    Auth.list_roles()
+    |> Enum.map(fn role -> {role.name, role.id} end)
+  end
 end
